Move transform dialect builder file to Common
Since we want to re-use functions for both CPU and GPU moving it all in
common for now. We should break it up once the API stabilise.
