+++ compiler/src/iree/compiler/Codegen/Common/TransformDialectStrategies.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Codegen/LLVMGPU/GPUTransformDialectStrategies.h"
+#include "iree/compiler/Codegen/Common/TransformDialectStrategies.h"
 
 #include <numeric>
 #include <type_traits>
@@ -15,7 +15,6 @@
 #include "iree/compiler/Codegen/Common/TransformExtensions/CommonExtensions.h"
 #include "iree/compiler/Codegen/Common/UserConfig.h"
 #include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
-#include "iree/compiler/Codegen/LLVMGPU/GPUTransformDialectStrategies.h"
 #include "iree/compiler/Codegen/LLVMGPU/KernelConfig.h"
 #include "iree/compiler/Codegen/LLVMGPU/TransformExtensions/LLVMGPUExtensions.h"
 #include "iree/compiler/Codegen/LLVMGPU/TransposeUtils.h"
